Dogs routes are fully functional, moved to dogs.js controller, doing the same for cats now
