+++ dogs.json
@@ -1,1 +1,1 @@
-[{"name":"Balto","image":"https://inland360.com/wp-content/uploads/2020/04/Trivia-Gunnar_Kaasen_with_Balto.jpg","famousFor":"saving a bunch of children by being the best sled doggo"},{"name":"Air Bud","image":"https://upload.wikimedia.org/wikipedia/en/7/70/Air_bud_world_pup2.jpg","famousFor":"For being an all-star athelete"}]
+[{"name":"Balto","image":"https://inland360.com/wp-content/uploads/2020/04/Trivia-Gunnar_Kaasen_with_Balto.jpg","famousFor":"saving a bunch of children by being the best sled doggo"},{"name":"Air Bud","image":"https://upload.wikimedia.org/wikipedia/en/7/70/Air_bud_world_pup2.jpg","famousFor":"For being an all-star athelete"},{"name":"Scooby Doo","image":"https://encrypted-tbn0.gstatic.com/images?q=tbn%3AANd9GcTKqWIuqh-O3FzlszBxBsuMJHfxIwud4FByAw&usqp=CAU","famousFor":"Solving mysteries with the Mystery Gang & always having the munchies"},{"name":"Courage The Cowardly Dog","image":"https://encrypted-tbn0.gstatic.com/images?q=tbn%3AANd9GcQsCMiaoyPJLQA0o8xBwn4GmavP_NPgqRGwAw&usqp=CAU","famousFor":"Saving Eustace and Muriel from the creepy things that happen in NoWhere"}]
