+++ views/dogs/show.ejs
@@ -1,3 +1,3 @@
-<h2> <%= dog.name %> is famous for: <%= dog.famousFor %>
-    <img src="<%= dog.image %>">  
+<h2> <%= myDog.name %> is famous for: <%= myDog.famousFor %>
+    <img src="<%= myDog.image %>">  
 </h2>
