+++ views/layout.ejs
@@ -8,7 +8,9 @@
 </head>
 <body>
     <ul>
-      
+        <li><a href="/dogs"> Dogs Home Page </a></li>
+        <li><a href="/dogs/new"> Add a Dog </a></li>
+
     </ul>
     <h1> DOGS AND CATS</h1>
     <%- body %>
